Change file storage, add login time limit
Files are now stored in ~/.yapm/.db.  Once a user logs in, symlinks
are created in it's current directory.  Those are removed once he
disconnects.

A user can now set a login time limit.  When a user logs in, it
creates a process with nohup which just sleeps and then call the
program with the disconnect option.  When a user disconnects manually,
it tries to find such process and kill it after disconnecting.